fix: avatar
