+++ src/app/router/routesConfig.tsx
@@ -1,5 +1,6 @@
+import ApplicationProgress from '../../features/TutorCabinet/ui/ApplicationProgress';
 import ProfileCard from '../../features/TutorCabinet/ui/ProfileCard';
-import { profileCardData } from '../../features/TutorCabinet/ui/ProfileCard/data';
+import { profileCardData } from '../../features/TutorCabinet/ui/ProfileCard/mockData';
 import MainPage from '../../pages/MainPage';
 import NotFoundPage from '../../pages/NotFoundPage';
 import TutorCatalogPage from '../../pages/TutorCatalogPage';
@@ -19,16 +20,22 @@ export const routesConfig: IRoute[] = [
   },
   {
     path: '/test',
-    // element: <ApplicationProgress isProfileHidden={true} onToggleVisibility={()=>{}}/>,
+
     element: (
-      <ProfileCard
-        avatar={profileCardData.avatar}
-        name={profileCardData.name}
-        status={profileCardData.status}
-        tg={profileCardData.tg}
-        link={profileCardData.link}
-        rating={profileCardData.rating}
-      />
+      <>
+        <ApplicationProgress
+          isProfileHidden={true}
+          onToggleVisibility={() => {}}
+        />
+        <ProfileCard
+          avatar={profileCardData.avatar}
+          name={profileCardData.name}
+          status={profileCardData.status}
+          tg={profileCardData.tg}
+          link={profileCardData.link}
+          rating={profileCardData.rating}
+        />
+      </>
     ),
     auth: false
   },
