+++ src/features/TutorCabinet/ui/ApplicationProgress/index.module.scss
@@ -27,14 +27,8 @@
   justify-content: space-between;
 
   &__text {
-  font-family: var(--ffm-default);
-  font-size: 14px;
-  line-height: 1.36;
+    font-family: var(--ffm-default);
+    font-size: 14px;
+    line-height: 1.36;
+  }
 }
-}
-
-
-
-
-
-
