+++ src/features/TutorCabinet/ui/ProfileCard/type.ts
@@ -1,0 +1,8 @@
+export interface ProfileCardProps {
+  name: string;
+  status: string;
+  tg: string;
+  link: string;
+  rating: number;
+  avatar: string;
+}
