+++ src/shared/api/index.ts
@@ -1,5 +1,4 @@
-const URL  = import.meta.env.VITE_REPETIT_ME_URL
-
+const URL = import.meta.env.VITE_REPETIT_ME_URL;
 
 export async function httpJson<T>(
   endpoint: string,
@@ -8,26 +7,26 @@ export async function httpJson<T>(
     headers?: Record<string, string>;
     body?: unknown;
     params?: Record<string, string | number | boolean | undefined>;
-  } 
-  ): Promise<T> {    
-    try {
-      const response = await fetch(`${URL}/${endpoint}`, {
-        method: options.method,
-        headers: {
-          'Content-Type': 'application/json',
-          ...options.headers,
-        },
-        body: options.body !== undefined ? JSON.stringify(options.body) : undefined,
-      });
-      
-      if (!response.ok) {
-        throw new Error(`Error, статус: ${response.status}`);
-      }
-      
-      const data = await response.json();
-      return data;
-    } 
-    catch (err) {
-      throw new Error(`Запрос не выполнен, ${err}`);
-    }
   }
+): Promise<T> {
+  try {
+    const response = await fetch(`${URL}/${endpoint}`, {
+      method: options.method,
+      headers: {
+        'Content-Type': 'application/json',
+        ...options.headers
+      },
+      body:
+        options.body !== undefined ? JSON.stringify(options.body) : undefined
+    });
+
+    if (!response.ok) {
+      throw new Error(`Error, статус: ${response.status}`);
+    }
+
+    const data = await response.json();
+    return data;
+  } catch (err) {
+    throw new Error(`Запрос не выполнен, ${err}`);
+  }
+}
